+++ route/delete_files_in_serv.js
@@ -12,28 +12,36 @@ router.route('/delete-in-serv')
 		var file = ""
 		var file_profileId = request.body.file_profileId
 		var pathImg = request.body.path
-		table.push(file_profileId)
+		ret.success = []
+		ret.global_msg = []
+		if (request.session.token == request.headers["x-access-token"]){
+			table.push(file_profileId)
 			if (pathImg){
 					file = pathImg.replace("/asset", "public")
 					setTimeout((function(file) {
-					 	return function(){
-					 		fs.access(file, fs.constants.F_OK, (err) => {
+						return function(){
+							fs.access(file, fs.constants.F_OK, (err) => {
 								if (!err){
 									fs.unlink(file, (err) => {
-									  if (err)
-									  	throw err
-									  console.log(file+' was deleted');
+									if (err)
+										throw err
+									console.log(file+' was deleted');
 									})
 								}
 							})
-					 	};
-					 }) (file), 100)
-					 ret.success = true
-					 response.send(ret)
+						};
+					}) (file), 100)
+					ret.success.push(true)
+					response.send(ret)
 			}
 			else{
-				ret.success = false
+				ret.success.push(false)
 				response.send(ret)
 			}
+		}else{
+			ret.success.push(false)
+			ret.global_msg.push("Session compromise !")
+			response.send(ret)
+		}
 	})
 module.exports = router
